requirements.txt
flask-accept==0.0.3
flask-bootstrap==3.3.7.1
flask-nav==0.6
flask==0.12.2
flask==0.12.2
gunicorn==19.7.1
pytest==3.0.2
psycopg2==2.6.1
Flask-SQLAlchemy===2.1
Flask-Migrate==1.8.0
